backend/requirements.txt
# PointX Backend Dependencies

flask
flask-cors
flask-smorest
marshmallow
supabase
postgrest
python-dotenv
bcrypt
PyJWT
qrcode[pil]
pillow
gunicorn
